% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_public_jinja.R
\name{compile_public_jinja}
\alias{compile_public_jinja}
\title{Takes in a codebook object and compiles the public jinja.}
\usage{
compile_public_jinja(active__codebook, stage, i)
}
\arguments{
\item{active__codebook:}{the active codebook to compile}

\item{stage:}{the stage of codebook being processed}

\item{i:}{dataset import object}
}
\value{
returns a dataframe object that has the public compiled from our pojrect dictionary.
}
\description{
Takes in a codebook object and compiles the public jinja.
}
\examples{
export_human_codebooks()
}
