% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_schema_from_linkage.R
\name{draw_schema_from_linkage}
\alias{draw_schema_from_linkage}
\title{Draws a schema diagram based on linkage metadata}
\usage{
draw_schema_from_linkage(df_linkage_tidy, dataset_id_tmp, i)
}
\arguments{
\item{df_linkage_tidy}{the schema metadata stored as a dataframe}

\item{dataset_id_tmp}{dataset id}

\item{i}{dataset sepcific improt object}
}
\value{
side effect. will draw schema diagram at the dataset's folde rin the repository
}
\description{
Takes in metadata from linkage.csv then autoamtatically draws a PNG formatted schematic.
}
\examples{
draw_schema_from_linkage()
}
