% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_human_codebooks.R
\name{export_human_codebooks}
\alias{export_human_codebooks}
\title{Exports human friendly codebook excell sheet}
\usage{
export_human_codebooks(df_denorm, df_data_details, i)
}
\arguments{
\item{df_denorm}{dernoamlized table}

\item{df_data_details}{supplemental details about data.}

\item{i}{dataset sepcific improt object}
}
\value{
side effect. will write human readable codebook.
}
\description{
WOrked wihtDMC to develop internal facing human accessible codebooks. THis is run in rpoduction for all fully renovated codebooks.
}
\examples{
export_human_codebooks()
}
