% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_server_path.R
\name{get_uhc_file_path}
\alias{get_uhc_file_path}
\title{Get the UHC server path for a certain file depending on the datasetid}
\usage{
get_uhc_file_path(dataset_id, file)
}
\arguments{
\item{dataset_id}{the folder name of your dataset in the FAIR Renovation folder.
For example `APSL1AD`.}

\item{file}{the name of file you want. possible options include any of the clean
files  (e.g  `var_name.csv`) or raw files (e.g. `APSL1AD_06132022.csv`)}
}
\value{
a path to file you specified
}
\description{
This function is useful for generating path names to the UHC server. It will only work if
have uploaded your files to the UHC server `\\files.drexel.edu\colleges\SOPH\Shared\UHC\Projects\Wellcome_Trust\Data Methods Core\Dashboards\FAIR Renovations`
}
\examples{
get_uhc_file_path(dataset_id = "APSL1AD", file = "APSL1AD_06132022.csv")
get_uhc_file_path(dataset_id = "APSL1AD", file = "var_name.csv")
}
