% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_salubral_xlsx_cdbk.R
\name{import_salurbal_xslx_cdbk}
\alias{import_salurbal_xslx_cdbk}
\title{Imports a SALURBAL formated excel codebook}
\usage{
import_salurbal_xslx_cdbk(xlsx_path, i = NULL, src = F)
}
\arguments{
\item{xlsx_path:}{the path to the .xslx to import}

\item{i:}{dataset import object}

\item{src:}{optinoal parapmeter to indicate if this is a source dictionary to import}
}
\value{
returns a named list of dataframes.
}
\description{
This will operationalize various sheets as a list of dataframes. These will serve as data
structures that will go through QC and ultimate dernoamlziationa nd export.
}
\examples{
import_salurbal_xslx_cdbk()
}
