% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_codebook_var.R
\name{sanitize_codebook_var}
\alias{sanitize_codebook_var}
\title{Convert codebook variable to generic format}
\usage{
sanitize_codebook_var(var_raw)
}
\arguments{
\item{var_raw}{the raw codebook variable. This often contains higher level information
such as year, gender or other attributes. We o}
}
\value{
A character value which is just the base variable name.
}
\description{
Convert codebook variable to generic format
}
\examples{
sanitize_codebook_var(var_raw = "OBESITY_L1AD_YYYY")
}
