% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot_excel.R
\name{snapshot_excel}
\alias{snapshot_excel}
\title{Creates a diff accessible snapshot of .xlsx file}
\usage{
snapshot_excel(
  dataset_id_tmp = NULL,
  path = NULL,
  quiet = T,
  folder = "datasets"
)
}
\arguments{
\item{dataset_id_tmp}{the dataset id of interests}

\item{path}{a specific path if desired}

\item{quiet}{boolean to indicate if CLI messages are to be generated}

\item{folder}{which folde rin the repositroy should we scan}
}
\value{
nothing, just side effects
}
\description{
this function will store codebooks in .xlsx in .json for a particular dataset.This will allow us to track changes between versions much better with git.
}
\examples{
snapshot_excel("DTH")
}
