% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack_year.R
\name{unpack_year}
\alias{unpack_year}
\title{Tidy year for input dataframe}
\usage{
unpack_year(df)
}
\arguments{
\item{df}{the input data frame}
}
\value{
returns dataframe
}
\description{
Takes in a dataframe and tidies or unpacks the year field. So any ranges or seperateotrs will be pivot longer.
}
\examples{
unpack_year(df = df_raw__codebook)
}
