% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_csv_utf8.R
\name{validate_csv_utf8}
\alias{validate_csv_utf8}
\title{Checks CSV for UTF8}
\usage{
validate_csv_utf8(csv_path)
}
\arguments{
\item{csv_path}{thepath to the csv to check}
}
\value{
returns boolean TRUE or FALSE
}
\description{
Takes in a CSV path  that CSV is fully UTF-8 encoded
}
\examples{
validate_csv_utf8(csv_path = "raw__codebook.csv")
}
